+++ irs-api/src/main/java/org/eclipse/tractusx/irs/semanticshub/SemanticsHubClient.java
@@ -1,9 +1,10 @@
 /********************************************************************************
- * Copyright (c) 2021,2022
- *       2022: Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
+ * Copyright (c) 2021,2022,2023
  *       2022: ZF Friedrichshafen AG
  *       2022: ISTOS GmbH
- * Copyright (c) 2021,2022 Contributors to the Eclipse Foundation
+ *       2022,2023: Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
+ *       2022,2023: BOSCH AG
+ * Copyright (c) 2021,2022,2023 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -26,18 +27,27 @@ import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.util.ArrayList;
 import java.util.Base64;
+import java.util.List;
 import java.util.Map;
 import java.util.Optional;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+import java.util.stream.Stream;
 
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.io.FilenameUtils;
 import org.apache.commons.lang3.StringUtils;
+import org.apache.commons.lang3.reflect.TypeUtils;
 import org.eclipse.tractusx.irs.configuration.RestTemplateConfig;
+import org.eclipse.tractusx.irs.configuration.SemanticsHubConfiguration;
 import org.eclipse.tractusx.irs.services.validation.SchemaNotFoundException;
 import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Profile;
+import org.springframework.core.ParameterizedTypeReference;
+import org.springframework.http.RequestEntity;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestClientException;
 import org.springframework.web.client.RestTemplate;
@@ -56,6 +66,7 @@ interface SemanticsHubClient {
      */
     String getModelJsonSchema(String urn) throws SchemaNotFoundException;
 
+    List<AspectModel> getAllAspectModels() throws SchemaNotFoundException;
 }
 
 /**
@@ -71,6 +82,15 @@ class SemanticsHubClientLocalStub implements SemanticsHubClient {
     public String getModelJsonSchema(final String urn) {
         return "{" + "  \"$schema\": \"http://json-schema.org/draft-07/schema#\"," + "  \"type\": \"integer\"" + "}";
     }
+
+    @Override
+    public List<AspectModel> getAllAspectModels() {
+        return List.of(
+                new AspectModel("urn:bamm:com.catenax.esr_certificates.esr_certificate:1.0.0#EsrCertificate", "1.0.0",
+                        "EsrCertificate", "BAMM", "RELEASED"),
+                new AspectModel("urn:bamm:io.catenax.assembly_part_relationship:1.0.0#AssemblyPartRelationship",
+                        "1.0.0", "AssemblyPartRelationship", "BAMM", "RELEASED"));
+    }
 }
 
 /**
@@ -81,23 +101,21 @@ class SemanticsHubClientLocalStub implements SemanticsHubClient {
 @Profile({ "!local && !test" })
 class SemanticsHubClientImpl implements SemanticsHubClient {
 
+    public static final String LOCAL_MODEL_TYPE = "BAMM";
+    public static final String LOCAL_MODEL_STATUS = "PROVIDED";
     private static final String PLACEHOLDER_URN = "urn";
-
+    private final SemanticsHubConfiguration config;
     private final RestTemplate restTemplate;
-    private final String semanticsHubUrl;
-    private final String localModelDirectory;
 
     /* package */ SemanticsHubClientImpl(
             @Qualifier(RestTemplateConfig.SEMHUB_REST_TEMPLATE) final RestTemplate restTemplate,
-            @Value("${semanticsHub.modelJsonSchemaEndpoint:}") final String semanticsHubUrl,
-            @Value("${semanticsHub.localModelDirectory:}") final String localModelDirectory) {
+            final SemanticsHubConfiguration config) {
+        this.config = config;
         this.restTemplate = restTemplate;
-        this.semanticsHubUrl = semanticsHubUrl;
-        this.localModelDirectory = localModelDirectory;
 
-        if (StringUtils.isNotBlank(semanticsHubUrl)) {
-            requirePlaceholder(semanticsHubUrl);
-        } else if (StringUtils.isBlank(localModelDirectory)) {
+        if (StringUtils.isNotBlank(config.getModelJsonSchemaEndpoint())) {
+            requirePlaceholder(config.getModelJsonSchemaEndpoint());
+        } else if (StringUtils.isBlank(config.getLocalModelDirectory())) {
             log.warn("No Semantic Hub URL or local model directory was provided. Cannot validate submodel payloads!");
         }
     }
@@ -121,9 +139,89 @@ class SemanticsHubClientImpl implements SemanticsHubClient {
                                                "Could not load model with URN " + urn));
     }
 
+    @Override
+    public List<AspectModel> getAllAspectModels() throws SchemaNotFoundException {
+        return readAllFromSemanticHub().or(this::readAllFromFilesystem).orElse(List.of());
+    }
+
+    private Optional<List<AspectModel>> readAllFromFilesystem() {
+        if (StringUtils.isNotBlank(config.getLocalModelDirectory())) {
+            final Path path = Paths.get(config.getLocalModelDirectory());
+            try (Stream<Path> stream = Files.list(path)) {
+                return Optional.of(stream.filter(file -> !Files.isDirectory(file))
+                                         .map(Path::getFileName)
+                                         .map(Path::toString)
+                                         .map(this::getDecodedString)
+                                         .map(this::createAspectModel)
+                                         .filter(Optional::isPresent)
+                                         .map(Optional::get)
+                                         .toList());
+            } catch (IOException e) {
+                log.error("Could not read schema Files.", e);
+            }
+        }
+        return Optional.empty();
+    }
+
+    private String getDecodedString(final String urnBase64) {
+        try {
+            return decode(urnBase64);
+        } catch (IllegalArgumentException e) {
+            log.error("Could not Base64 decode urn.", e);
+            return urnBase64;
+        }
+    }
+
+    private Optional<AspectModel> createAspectModel(final String urn) {
+        log.debug("Extracting aspect information for urn: '{}'", urn);
+        final Matcher matcher = Pattern.compile("^urn:bamm:.*:(\\d\\.\\d\\.\\d)#(\\w+)$").matcher(urn);
+        if (matcher.find()) {
+            final String version = matcher.group(1);
+            final String name = matcher.group(2);
+            return Optional.of(new AspectModel(urn, version, name, LOCAL_MODEL_TYPE, LOCAL_MODEL_STATUS));
+        }
+        log.warn("Could not extract aspect information from urn: '{}'", urn);
+        return Optional.empty();
+    }
+
+    private Optional<List<AspectModel>> readAllFromSemanticHub() {
+        log.info("Reading models from semantic hub.");
+        if (StringUtils.isNotBlank(config.getUrl())) {
+            int currentPage = 0;
+            final List<AspectModel> aspectModelsCollection = new ArrayList<>();
+            Optional<PaginatedResponse<AspectModel>> semanticHubPage;
+            do {
+                semanticHubPage = getSemanticHubPage(currentPage++, config.getPageSize());
+                log.info("Got response from semantic hub '{}'", semanticHubPage.toString());
+                aspectModelsCollection.addAll(semanticHubPage.orElseThrow().getContent());
+            } while (semanticHubPage.isPresent() && semanticHubPage.get().hasNext());
+
+            return Optional.of(aspectModelsCollection);
+        }
+        return Optional.empty();
+    }
+
+    private Optional<PaginatedResponse<AspectModel>> getSemanticHubPage(final int page, final int pageSize) {
+        try {
+            log.info("Request semantic hub page '{}'  with size '{}' for url '{}'", page, pageSize, config.getUrl());
+            final UriComponentsBuilder uriBuilder = UriComponentsBuilder.fromUriString(config.getUrl())
+                                                                        .queryParam("page", page)
+                                                                        .queryParam("pageSize", pageSize);
+            log.info("Semantic Hub URL '{}'", uriBuilder.toUriString());
+            final ParameterizedTypeReference<PaginatedResponse<AspectModel>> responseType = ParameterizedTypeReference.forType(
+                    TypeUtils.parameterize(PaginatedResponse.class, AspectModel.class));
+            final ResponseEntity<PaginatedResponse<AspectModel>> result = restTemplate.exchange(
+                    RequestEntity.get(uriBuilder.toUriString()).build(), responseType);
+            return Optional.ofNullable(result.getBody());
+        } catch (RestClientException e) {
+            log.error("Unable to retrieve models from semantic hub.", e);
+        }
+        return Optional.empty();
+    }
+
     private Optional<String> readFromFilesystem(final String urn) {
-        if (StringUtils.isNotBlank(localModelDirectory)) {
-            final Path path = Paths.get(localModelDirectory, normalize(urn));
+        if (StringUtils.isNotBlank(config.getLocalModelDirectory())) {
+            final Path path = Paths.get(config.getLocalModelDirectory(), normalize(urn));
             if (path.toFile().exists()) {
                 try {
                     return Optional.of(Files.readString(path));
@@ -136,9 +234,10 @@ class SemanticsHubClientImpl implements SemanticsHubClient {
     }
 
     private Optional<String> readFromSemanticHub(final String urn) {
-        if (StringUtils.isNotBlank(semanticsHubUrl)) {
+        if (StringUtils.isNotBlank(config.getModelJsonSchemaEndpoint())) {
             try {
-                final UriComponentsBuilder uriBuilder = UriComponentsBuilder.fromUriString(semanticsHubUrl);
+                final UriComponentsBuilder uriBuilder = UriComponentsBuilder.fromUriString(
+                        config.getModelJsonSchemaEndpoint());
                 final Map<String, String> values = Map.of(PLACEHOLDER_URN, urn);
                 return Optional.ofNullable(restTemplate.getForObject(uriBuilder.build(values), String.class));
             } catch (final RestClientException e) {
@@ -150,5 +249,9 @@ class SemanticsHubClientImpl implements SemanticsHubClient {
 
     private String normalize(final String urn) {
         return Base64.getEncoder().encodeToString(FilenameUtils.getName(urn).getBytes(StandardCharsets.UTF_8));
+    }
+
+    private String decode(final String urnBase64) {
+        return new String(Base64.getDecoder().decode(urnBase64), StandardCharsets.UTF_8);
     }
 }