+++ irs-api/src/main/java/org/eclipse/tractusx/irs/semanticshub/SemanticsHubFacade.java
@@ -1,9 +1,10 @@
 /********************************************************************************
- * Copyright (c) 2021,2022
- *       2022: Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
+ * Copyright (c) 2021,2022,2023
  *       2022: ZF Friedrichshafen AG
  *       2022: ISTOS GmbH
- * Copyright (c) 2021,2022 Contributors to the Eclipse Foundation
+ *       2022,2023: Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
+ *       2022,2023: BOSCH AG
+ * Copyright (c) 2021,2022,2023 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -21,6 +22,10 @@
  ********************************************************************************/
 package org.eclipse.tractusx.irs.semanticshub;
 
+import java.time.Clock;
+import java.time.ZonedDateTime;
+import java.util.List;
+
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.eclipse.tractusx.irs.services.validation.SchemaNotFoundException;
@@ -37,6 +42,7 @@ import org.springframework.stereotype.Service;
 public class SemanticsHubFacade {
 
     private static final String SEMANTICS_HUB_CACHE_NAME = "schema_cache";
+    private static final String SEMANTICS_HUB_MODEL_CACHE_NAME = "model_cache";
 
     private final SemanticsHubClient semanticsHubClient;
 
@@ -57,6 +63,18 @@ public class SemanticsHubFacade {
     @CacheEvict(value = SEMANTICS_HUB_CACHE_NAME, allEntries = true)
     public void evictAllCacheValues() {
         log.debug("Clearing Semantics Hub Cache.");
+    }
+
+    /**
+     * Search in Semantic hub or local models for all available aspect models.
+     * @return All available aspect models.
+     */
+    @Cacheable(SEMANTICS_HUB_MODEL_CACHE_NAME)
+    public AspectModels getAllAspectModels() throws SchemaNotFoundException {
+        final List<AspectModel> aspectModels = this.semanticsHubClient.getAllAspectModels();
+
+        final String lastUpdated = ZonedDateTime.now(Clock.systemUTC()).toString();
+        return AspectModels.builder().models(aspectModels).lastUpdated(lastUpdated).build();
     }
 
 }