+++ irs-api/src/test/java/org/eclipse/tractusx/irs/controllers/IrsExceptionHandlerTest.java
@@ -1,9 +1,10 @@
 /********************************************************************************
- * Copyright (c) 2021,2022
- *       2022: Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
+ * Copyright (c) 2021,2022,2023
  *       2022: ZF Friedrichshafen AG
  *       2022: ISTOS GmbH
- * Copyright (c) 2021,2022 Contributors to the Eclipse Foundation
+ *       2022,2023: Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
+ *       2022,2023: BOSCH AG
+ * Copyright (c) 2021,2022,2023 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -24,6 +25,7 @@ package org.eclipse.tractusx.irs.controllers;
 import static org.eclipse.tractusx.irs.util.TestMother.registerJobWithoutDepthAndAspect;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 import static org.springframework.web.client.HttpServerErrorException.InternalServerError;
@@ -31,14 +33,17 @@ import static org.springframework.web.client.HttpServerErrorException.InternalSe
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.eclipse.tractusx.irs.configuration.SecurityConfiguration;
 import org.eclipse.tractusx.irs.services.IrsItemGraphQueryService;
+import org.eclipse.tractusx.irs.services.SemanticHubService;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.context.annotation.Import;
 import org.springframework.http.MediaType;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.test.context.support.WithMockUser;
 import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.web.method.annotation.MethodArgumentTypeMismatchException;
 
 @WebMvcTest(IrsController.class)
 @Import(SecurityConfiguration.class)
@@ -46,6 +51,8 @@ class IrsExceptionHandlerTest {
 
     @MockBean
     private IrsItemGraphQueryService service;
+    @MockBean
+    private SemanticHubService semanticHubService;
 
     @Autowired
     private MockMvc mockMvc;
@@ -59,5 +66,62 @@ class IrsExceptionHandlerTest {
                                               .content(new ObjectMapper().writeValueAsString(
                                                       registerJobWithoutDepthAndAspect())))
                     .andExpect(status().is5xxServerError());
+    }
+
+    @Test
+    @WithMockUser(authorities = "view_irs")
+    void shouldReturn500WhenGetSemanticModelsFails() throws Exception {
+        when(semanticHubService.getAllAspectModels()).thenThrow(InternalServerError.class);
+
+        this.mockMvc.perform(get("/irs/aspectmodels").contentType(MediaType.APPLICATION_JSON)
+                                                     .content(new ObjectMapper().writeValueAsString(
+                                                             registerJobWithoutDepthAndAspect())))
+                    .andExpect(status().is5xxServerError());
+    }
+
+    @Test
+    @WithMockUser(authorities = "view_irs")
+    void shouldReturn400WhenProvidingBadInput() throws Exception {
+        when(semanticHubService.getAllAspectModels()).thenThrow(IllegalArgumentException.class);
+
+        this.mockMvc.perform(get("/irs/aspectmodels").contentType(MediaType.APPLICATION_JSON)
+                                                     .content(new ObjectMapper().writeValueAsString(
+                                                             registerJobWithoutDepthAndAspect())))
+                    .andExpect(status().isBadRequest());
+    }
+
+    @Test
+    @WithMockUser(authorities = "view_irs")
+    void shouldReturn400WhenCatchingIllegalStateException() throws Exception {
+        when(semanticHubService.getAllAspectModels()).thenThrow(IllegalStateException.class);
+
+        this.mockMvc.perform(get("/irs/aspectmodels").contentType(MediaType.APPLICATION_JSON)
+                                                     .content(new ObjectMapper().writeValueAsString(
+                                                             registerJobWithoutDepthAndAspect())))
+                    .andExpect(status().isBadRequest());
+    }
+
+
+    @Test
+    @WithMockUser(authorities = "view_irs")
+    void shouldReturn400WhenCatchingMethodArgumentTypeMismatchException() throws Exception {
+        when(semanticHubService.getAllAspectModels()).thenThrow(MethodArgumentTypeMismatchException.class);
+
+        this.mockMvc.perform(get("/irs/aspectmodels").contentType(MediaType.APPLICATION_JSON)
+                                                     .content(new ObjectMapper().writeValueAsString(
+                                                             registerJobWithoutDepthAndAspect())))
+                    .andExpect(status().isBadRequest());
+    }
+
+
+    @Test
+    @WithMockUser(authorities = "view_irs")
+    void shouldReturn403WhenRightsAreMissing() throws Exception {
+        when(semanticHubService.getAllAspectModels()).thenThrow(AccessDeniedException.class);
+
+        this.mockMvc.perform(get("/irs/aspectmodels").contentType(MediaType.APPLICATION_JSON)
+                                                     .content(new ObjectMapper().writeValueAsString(
+                                                             registerJobWithoutDepthAndAspect())))
+                    .andExpect(status().isForbidden());
     }
 }