+++ irs-api/src/test/java/org/eclipse/tractusx/irs/services/OutboundMeterRegistryServiceTest.java
@@ -1,9 +1,10 @@
 /********************************************************************************
- * Copyright (c) 2021,2022
- *       2022: Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
+ * Copyright (c) 2021,2022,2023
  *       2022: ZF Friedrichshafen AG
  *       2022: ISTOS GmbH
- * Copyright (c) 2021,2022 Contributors to the Eclipse Foundation
+ *       2022,2023: Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
+ *       2022,2023: BOSCH AG
+ * Copyright (c) 2021,2022,2023 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -26,12 +27,12 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
+import java.util.Set;
+
 import io.github.resilience4j.retry.Retry;
 import io.github.resilience4j.retry.RetryRegistry;
 import io.micrometer.core.instrument.MeterRegistry;
 import io.micrometer.core.instrument.simple.SimpleMeterRegistry;
-import io.vavr.collection.Array;
-import io.vavr.collection.Seq;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -54,7 +55,7 @@ class OutboundMeterRegistryServiceTest {
         final Retry.EventPublisher publisher = mock(Retry.EventPublisher.class);
         when(retry.getEventPublisher()).thenReturn(publisher);
 
-        final Seq<Retry> seq = Array.empty();
+        final Set<Retry> seq = Set.of();
         when(retryRegistry.getAllRetries()).thenReturn(seq);
         testee = new OutboundMeterRegistryService(meterRegistry, retryRegistry);